+++ src/entities/album/model/types.ts
