+++ src/entities/photo/index.ts
@@ -1,4 +1,4 @@
-export type { Photo } from "./model/photo";
+export type { Photo } from "./model/types";
 export { PhotoCard } from "./ui/PhotoCard";
 export { useGetPhotoByIdQuery, useGetPhotosByAlbumIdQuery, useGetAllPhotosQuery, useRefreshUserPhotosMutation } from "./api/photosApi";
 export { photosApi } from "./api/photosApi";
