+++ src/entities/todo/index.ts
@@ -1,4 +1,4 @@
-export type { Todo } from "./model/todo";
+export type { Todo } from "./model/types";
 export { TodoCard } from "./ui/TodoCard";
 export { useGetAllTodosQuery, useGetTodoByIdQuery, useGetTodosByUserIdQuery, useRefreshUserTodosMutation } from "./api/todosApi";
 export { todosApi } from "./api/todosApi";
