+++ src/entities/user/api/usersApi.ts
@@ -1,5 +1,5 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
-import type { User } from '../model/user'
+import type { User } from '../model/types'
 
 export const usersApi = createApi({
   reducerPath: 'usersApi',
