+++ src/entities/user/index.ts
@@ -1,4 +1,4 @@
-export type { User } from "./model/user";
+export type { User } from "./model/types";
 export { useGetUserByIdQuery, useGetAllUsersQuery } from "./api/usersApi";
 export { usersReducer, selectAllUsers, selectUserIds } from "./slice/usersSlice";
 export { setSelectedUserId } from "./slice/usersSlice";
